refactor: improve file system and fix all left eslint issues
