+++ src/components/PageProducts/conponents/Toolbar/Toolbar.ts
@@ -1,9 +1,8 @@
-import { space } from '../../../../const/store-name';
+import { route } from '../../../../router/route';
 import { createElem } from '../../../../utils/create-element';
 import { createLink } from '../../../../utils/create-link-element';
 import { newNameProduct } from '../../../../utils/edit-name-products';
 import styles from './Toolbar.module.scss';
-import { router } from '../../../../utils/router';
 
 export const renderProductPageToolbar = (product: ExtendedProduct): HTMLElement => {
   const toolbarProdPage: HTMLElement = createElem('div', styles['toolbar__product-page']);
@@ -14,12 +13,12 @@ export const renderProductPageToolbar = (product: ExtendedProduct): HTMLElement 
   toolbarProdPage.append(mainLink);
   mainLink.onclick = (e: Event): void => {
     e.preventDefault();
-    router('/');
+    route('/');
   };
 
   path.forEach((elem) => {
     const itemScope = createElem('div', styles['product-page__item-score']);
-    itemScope.innerHTML = `${space}/ ${elem}`;
+    itemScope.innerHTML = `&nbsp;/ ${elem}`;
     toolbarProdPage.append(itemScope);
   });
