+++ src/const/store-name.ts
@@ -1,2 +1,1 @@
 export const storeName = 'markett';
-export const space = '&nbsp;';
